fix: move legacy unit tests, fix route functions
