+++ src/applications/personalization/profile/routes.js
@@ -4,10 +4,12 @@ import { getRoutesForNav } from './routesForNav';
 import { Hub } from './components/hub/Hub';
 
 // conditionally add the profile hub route based on feature toggle
-const getRoutes = ({
-  profileContacts = false,
-  profileShowDirectDepositSingleForm = false,
-}) => {
+const getRoutes = (
+  { profileContacts = false, profileShowDirectDepositSingleForm = false } = {
+    profileContacts: false,
+    profileShowDirectDepositSingleForm: false,
+  },
+) => {
   return [
     ...getRoutesForNav({
       profileContacts,
